fix: solved issues in initiative department calculation
