+++ src/Core/Interfaces/Repositories/Locations/ILocationPolygonRepository.cs
@@ -5,10 +5,6 @@ using System.Threading.Tasks;
 
 using IAVH.BioTablero.CM.Core.Domain.Entities.Geo;
 
-using NetTopologySuite.Geometries;
-
-using LocationCustom = IAVH.BioTablero.CM.Core.Domain.Entities.Geo.Location;
-
 /// <summary>
 /// Location Polygon repository interface.
 /// </summary>
@@ -21,12 +17,4 @@ public interface ILocationPolygonRepository : IRepository<LocationPolygon, int>
     /// <param name="ct">Cancellation token.</param>
     /// <returns>Selected simplified polygon.</returns>
     Task<string> GetSimplifiedByLocationAsync(int locationId, CancellationToken ct = default);
-
-    /// <summary>
-    /// Get department by coordinate.
-    /// </summary>
-    /// <param name="coordinate">Geographic coordinate.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>The department to wich the coordinate belongs. Null if doesn't belong to any department.</returns>
-    Task<LocationCustom> GetDepartmentByCoordinateAsync(Point coordinate, CancellationToken ct = default);
 }
